requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
tensorflow>=2.8.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
websockets>=10.0
pydantic>=1.8.0
joblib

# Database dependencies
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
alembic>=1.7.0

# Environment and utilities
python-dotenv>=0.19.0
python-multipart>=0.0.5

# Optional: For production
redis>=4.0.0
celery>=5.2.0
